chore: rename InputLabel to TextLabel
