+++ packages/react-styled-core/src/TextLabel/index.js
@@ -1,0 +1,28 @@
+import React, { forwardRef } from 'react';
+import Text from '../Text';
+import useColorMode from '../useColorMode';
+
+const TextLabel = forwardRef((props, ref) => {
+  const { colorMode } = useColorMode();
+  const colorProps = {
+    dark: {
+      color: 'white:secondary',
+    },
+    light: {
+      color: 'black:secondary',
+    },
+  }[colorMode];
+
+  return (
+    <Text
+      as="label"
+      ref={ref}
+      {...colorProps}
+      {...props}
+    />
+  );
+});
+
+TextLabel.displayName = 'TextLabel';
+
+export default TextLabel;
