+++ packages/styled-docs/pages/textlabel.mdx
@@ -1,0 +1,27 @@
+# TextLabel
+
+The `TextLabel` component is a label that composes the `Text` component.
+
+## Import
+
+```js
+import TextLabel from '@trendmicro/react-styled-core/TextLabel';
+// or
+import { TextLabel } from '@trendmicro/react-styled-core';
+```
+
+## Usage
+
+```jsx
+<TextLabel>Label text:</TextLabel>
+```
+
+The text colors in the dark mode and light mode are `white:secondary` and `black:secondary`, respectively, defined in `theme.colors`.
+
+## Props
+
+`TextLabel` composes the [`Text`](./text) component. You can override the default styles with style props.
+
+| Name | Type | Default | Description |
+| :--- | :--- | :------ | :---------- |
+| `size` | `string` | | One of: '4xl', '3xl', '2xl', 'xl', 'lg', 'md', 'sm', 'xs' |
